feat: package model data + docker image
Summary: Adds a docker image with pre-packed model data. We use the new impure
derivations feature to run the embedding server on its own and download data,
"purify" it with a fixed-output derivation, and then use that as an input to
the actual server derivation that runs the hypercorn server.

In order to do this we add new tools to download data/load data from disk but
these are transparent to uses of 'nix run' or 'nix build'. Running the script
manually will use an on-disk cache under $HOME (as usual with pytorch), which
is fine for development.

This also adds a preliminary CI step to build the server, and publish it too!

Test Plan: build docker image, docker load, tested endpoint with httpie